requirements-api.txt
# Core dependencies
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5,<1.0.0
pydantic>=1.8.0,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=0.19.0,<1.0.0

# Search and NLP
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0
spacy>=3.2.0,<4.0.0
sentence-transformers>=2.2.0,<3.0.0
chromadb>=0.3.0,<1.0.0

# Development
debugpy>=1.5.0,<2.0.0
pytest>=6.2.5,<7.0.0
httpx>=0.23.0,<1.0.0
pylint>=2.12.0,<3.0.0
black>=22.3.0,<23.0.0
isort>=5.10.0,<6.0.0
